+++ libs/loader.sh
@@ -5,3 +5,4 @@
 
 source $DIR/libs/utilities.sh
 source $DIR/libs/docker.sh
+source $DIR/libs/git.sh