+++ libs/loader.sh
@@ -3,6 +3,9 @@
 # loader.sh
 #
 
-source $DIR/libs/utilities.sh
-source $DIR/libs/docker.sh
-source $DIR/libs/git.sh
+libraries=$( ls "$DIR/libs/" )
+for library in $libraries; do
+    if [[ "$library" != "loader.sh" ]]; then
+        source "$DIR/libs/$library"
+    fi
+done